backend/requirements.txt
coverage
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
gunicorn
django-environ
psycopg2-binary
